requirements.txt
anyio==4.12.0
asgiref==3.11.0
async-timeout==5.0.1
certifi==2025.11.12
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.4
Django==5.2.9
django-cors-headers==4.9.0
djangorestframework==3.16.1
exceptiongroup==1.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
msgpack==1.1.2
psycopg2-binary==2.9.11
python-dotenv==1.2.1
redis==7.1.0
requests==2.32.5
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
gunicorn
